worked on game entities and frameworke.py

+++ Server/src/game/post_data_interfaces/IAbility.py
@@ -9,8 +9,8 @@ class IAbility(IEntity):
     behavior: int
     abilityType: int
     maxLevel: int
-    cooldownTimeRemaining: int
-    cooldownTime: int
+    cooldownTimeRemaining: float
+    cooldownTime: float
     targetType: int
     abilityIndex: int
     type: str
